+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Removed deprecated discipline helpers sails.helpers.discipline.ban* [Issue 64](https://github.com/Lovinity/wwsu/issues/64)
  - Removed deprecated queue parameter in sails.helpers.rest.cmd [Issue 60](https://github.com/Lovinity/wwsu/issues/60)
  - Removed deprecated Songs.pendingCmd [Issue 59](https://github.com/Lovinity/wwsu/issues/59)
+ - Removed resume parameter from sails.helpers.playlists.start [Issue 58](https://github.com/Lovinity/wwsu/issues/58)
 
 ### Added
  - "device" property returned for recipient objects in recipients/get. This will contain the OneSignal UUID if the recipient is using the WWSU mobile app. Otherwise, it will be null.